Separate data-view-right (with some part of post edit) and collection-toggle
