+++ legacy/app/main/data/collection-toggle/collection-toggle-button.html
