+++ legacy/app/main/data/data-view-right/post-detail/add-form.html
