+++ legacy/app/main/data/data-view-right/post-detail/map.html
