+++ legacy/app/main/data/data-view-right/post-detail/post-category-value.directive.js
