+++ legacy/app/main/data/data-view-right/post-detail/post-detail-data.directive.js
