+++ legacy/app/main/data/data-view-right/post-detail/post-lock.service.js
