+++ legacy/app/main/data/data-view-right/post-detail/post-media-value.html
