+++ legacy/app/main/data/data-view-right/post-detail/post-value.directive.js
