+++ legacy/app/main/data/data-view-right/post-detail/post-video-value.directive.js
