+++ legacy/app/main/data/data-view-right/post-edit/post-data-editor.directive.js
