+++ legacy/app/main/data/data-view-right/post-edit/post-toolbox.directive.js
